Adding proper logging in all functions && cleaning up functions && updating sawmill
