+++ cmd/appleauth/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.7: Adding more logs & separating out types and helpers
 - v1.0.0-beta.6: Fixing merge
 - v1.0.0-beta.5: Reving Fireabse package to the latest
 - v1.0.0-beta.4: Removing GetAppleDeveloper token to Firebase pkg
