+++ cmd/appleauth/appleauth.go
@@ -1,11 +1,8 @@
 package appleauth
 
 import (
-	"context"
-	"fmt"
 	"log"
 	"net/http"
-	"os"
 
 	"cloud.google.com/go/firestore"
 	"github.com/pixelogicdev/gruveebackend/pkg/firebase"
@@ -14,19 +11,16 @@ import (
 )
 
 // zebcode - "Zebcode Rules 🦸‍♂️" (04/29/20)
-type appleDevTokenResp struct {
-	Token string
-}
-
-var firestoreClient *firestore.Client
-var logger sawmill.Logger
-var appleDevToken firebase.FirestoreAppleDevJWT
-var httpClient *http.Client
-var hostname string
-var templatePath string
+var (
+	firestoreClient *firestore.Client
+	logger          sawmill.Logger
+	appleDevToken   firebase.FirestoreAppleDevJWT
+	httpClient      *http.Client
+	hostname        string
+	templatePath    string
+)
 
 func init() {
-	httpClient = &http.Client{}
 	log.Println("AuthorizeWithApple initialized.")
 }
 
@@ -40,14 +34,15 @@ func AuthorizeWithApple(writer http.ResponseWriter, request *http.Request) {
 		return
 	}
 
-	// Get Apple Dev Token
 	// DR_DinoMight: Dammmmn, Apple Really?!?!?! (08/11/20)
 	appleDevToken, appleDevTokenErr := firebase.GetAppleDeveloperToken()
 	if appleDevTokenErr != nil {
 		http.Error(writer, appleDevTokenErr.Error(), http.StatusInternalServerError)
-		log.Printf("[GetAppleMusicMedia]: %v", appleDevTokenErr)
+		logger.LogErr("GetAppleMusicMedia", appleDevTokenErr, nil)
 		return
 	}
+
+	logger.Log("GetAppleDeveloperToken", "AppleDevToken recieved.")
 
 	// Render template
 	render := render.New(render.Options{
@@ -59,41 +54,6 @@ func AuthorizeWithApple(writer http.ResponseWriter, request *http.Request) {
 		logger.LogErr("Render", renderErr, nil)
 		return
 	}
-}
 
-// Helpers
-// initWithEnv takes our yaml env variables and maps them properly.
-// Unfortunately, we had to do this is main because in init we weren't able to access env variables
-func initWithEnv() error {
-	if os.Getenv("APPLE_TEAM_ID") == "" {
-		return fmt.Errorf("authorizeWithApple - APPLE_TEAM_ID does not exist")
-	}
-
-	// Get paths
-	var currentProject string
-	if os.Getenv("ENVIRONMENT") == "DEV" {
-		currentProject = os.Getenv("FIREBASE_PROJECTID_DEV")
-		hostname = os.Getenv("HOSTNAME_DEV")
-		templatePath = os.Getenv("APPLE_AUTH_TEMPLATE_PATH_DEV")
-	} else if os.Getenv("ENVIRONMENT") == "PROD" {
-		currentProject = os.Getenv("FIREBASE_PROJECTID_PROD")
-		hostname = os.Getenv("HOSTNAME_PROD")
-		templatePath = os.Getenv("APPLE_AUTH_TEMPLATE_PATH_PROD")
-	}
-
-	// Initialize Firestore
-	client, err := firestore.NewClient(context.Background(), currentProject)
-	if err != nil {
-		return fmt.Errorf("AuthorizeWithApple [Init Firestore]: %v", err)
-	}
-
-	// Initialize Sawmill
-	sawmillLogger, err := sawmill.InitClient(currentProject, os.Getenv("GCLOUD_CONFIG"), os.Getenv("ENVIRONMENT"), "AuthorizeWithApple")
-	if err != nil {
-		log.Printf("AuthorizeWithApple [Init Sawmill]: %v", err)
-	}
-
-	firestoreClient = client
-	logger = sawmillLogger
-	return nil
+	writer.WriteHeader(http.StatusOK)
 }
