+++ cmd/appleauth/go.mod
@@ -6,6 +6,7 @@ require (
 	cloud.google.com/go/firestore v1.2.0
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/pixelogicdev/gruveebackend/pkg/firebase v1.0.0-beta.13
+	github.com/pixelogicdev/gruveebackend/pkg/sawmill v1.0.0-beta.1
 	github.com/square/go-jose v2.5.1+incompatible
 	github.com/unrolled/render v1.0.3
 )
