+++ cmd/appleauth/helpers.go
@@ -1,0 +1,51 @@
+package appleauth
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"net/http"
+	"os"
+
+	"cloud.google.com/go/firestore"
+	"github.com/pixelogicdev/gruveebackend/pkg/sawmill"
+)
+
+// initWithEnv takes our yaml env variables and maps them properly.
+// Unfortunately, we had to do this is main because in init we weren't able to access env variables
+func initWithEnv() error {
+	if os.Getenv("APPLE_TEAM_ID") == "" {
+		return fmt.Errorf("authorizeWithApple - APPLE_TEAM_ID does not exist")
+	}
+
+	// Get paths
+	var currentProject string
+	if os.Getenv("ENVIRONMENT") == "DEV" {
+		currentProject = os.Getenv("FIREBASE_PROJECTID_DEV")
+		hostname = os.Getenv("HOSTNAME_DEV")
+		templatePath = os.Getenv("APPLE_AUTH_TEMPLATE_PATH_DEV")
+	} else if os.Getenv("ENVIRONMENT") == "PROD" {
+		currentProject = os.Getenv("FIREBASE_PROJECTID_PROD")
+		hostname = os.Getenv("HOSTNAME_PROD")
+		templatePath = os.Getenv("APPLE_AUTH_TEMPLATE_PATH_PROD")
+	}
+
+	// Init HTTP Client
+	httpClient = &http.Client{}
+
+	// Initialize Firestore
+	client, err := firestore.NewClient(context.Background(), currentProject)
+	if err != nil {
+		return fmt.Errorf("AuthorizeWithApple [Init Firestore]: %v", err)
+	}
+
+	// Initialize Sawmill
+	sawmillLogger, err := sawmill.InitClient(currentProject, os.Getenv("GCLOUD_CONFIG"), os.Getenv("ENVIRONMENT"), "AuthorizeWithApple")
+	if err != nil {
+		log.Printf("AuthorizeWithApple [Init Sawmill]: %v", err)
+	}
+
+	firestoreClient = client
+	logger = sawmillLogger
+	return nil
+}
