+++ cmd/appleauth/types.go
@@ -1,0 +1,6 @@
+package appleauth
+
+// appleDevTokenResp in the token that comes back from a successful Apple Music Auth call
+type appleDevTokenResp struct {
+	Token string
+}
