+++ cmd/createprovideruser/.version
@@ -1,2 +1,3 @@
+- v1.0.0-beta.3: Adding more log statements & creating helpers and types files
 - v1.0.0-beta.2: Fixing merge
 - v1.0.0-beta.1: Initial instance of function
