+++ cmd/createprovideruser/helpers.go
@@ -1,0 +1,41 @@
+package createprovideruser
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+
+	"cloud.google.com/go/firestore"
+	"github.com/pixelogicdev/gruveebackend/pkg/sawmill"
+)
+
+// initWithEnv takes our yaml env variables and maps them properly.
+// Unfortunately, we had to do this is main because in init we weren't able to access env variables
+func initWithEnv() error {
+	// Get paths
+	var currentProject string
+
+	// Get Project ID
+	if os.Getenv("ENVIRONMENT") == "DEV" {
+		currentProject = os.Getenv("FIREBASE_PROJECTID_DEV")
+	} else if os.Getenv("ENVIRONMENT") == "PROD" {
+		currentProject = os.Getenv("FIREBASE_PROJECTID_PROD")
+	}
+
+	// Initialize Firestore
+	client, err := firestore.NewClient(context.Background(), currentProject)
+	if err != nil {
+		return fmt.Errorf("CreateProviderUser [Init Firestore]: %v", err)
+	}
+
+	// Initialize Sawmill
+	sawmillLogger, err := sawmill.InitClient(currentProject, os.Getenv("GCLOUD_CONFIG"), os.Getenv("ENVIRONMENT"), "CreateProviderUser")
+	if err != nil {
+		log.Printf("CreateAppleDevToken [Init Sawmill]: %v", err)
+	}
+
+	firestoreClient = client
+	logger = sawmillLogger
+	return nil
+}
