+++ cmd/createsocialplaylist/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.9: Adding more logging & creating types and helpers files
 - v1.0.0-beta.8: Fixing merge
 - v1.0.0-beta.7: Reving firebase pkg and social pkg
 - v1.0.0-beta.6: Adding support for Apple Music Playlist creation
