+++ cmd/createuser/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.5: Adding more logging & adding helpers file
 - v1.0.0-beta.4: Fixing merge
 - v1.0.0-beta.3: Reving firebase package to latest
 - v1.0.0-beta.2: Changing enviornment variable check location
