+++ cmd/fetchallmedia/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.4: Adding better logging
 - v1.0.0-beta.3: Fixing merge
 - v1.0.0-beta.2: Reving Firebase pkg to the latest
 - v1.0.0-beta.1: Initializing function
