+++ cmd/fetchallmedia/fetchallmedia.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"log"
 	"strings"
 
 	"github.com/pixelogicdev/gruveebackend/pkg/firebase"
@@ -17,12 +16,13 @@ var (
 
 // FetchAllMedia queries all music providers for songs data when a new document is added
 func FetchAllMedia(ctx context.Context, event firebase.FirestoreEventSongs) error {
+	logger.Log("FetchAllMedia", fmt.Sprintf("Event received: %v", event))
+
 	initErr := initWithEnv()
 	if initErr != nil {
 		logger.LogErr("InitWithErr", initErr, nil)
 		return initErr
 	}
-	log.Println(event)
 
 	// Init data
 	var appleData *map[string]interface{}
@@ -35,23 +35,28 @@ func FetchAllMedia(ctx context.Context, event firebase.FirestoreEventSongs) erro
 	mediaType := event.Value.Fields.Type.StringValue
 	docPath := strings.Split(event.Value.Name, "documents/")
 
+	logger.Log("FetchAllMedia", fmt.Sprintf("MediaName: %v", mediaName))
+	logger.Log("FetchAllMedia", fmt.Sprintf("MediaCreator: %v", mediaCreator))
+	logger.Log("FetchAllMedia", fmt.Sprintf("MediaType: %v", mediaType))
+	logger.Log("FetchAllMedia", fmt.Sprintf("DocPath: %v", docPath))
+
 	// Get media name and creator
 	// This can be a album or playlist or song
 	// We 1000% cannot get the playlist in another platform (until we start mapping the songs from a playlist)
 	if mediaType == "playlist" {
-		log.Println("[FetchAllMedia] Media is a playlist, we don't need to run a check.")
+		logger.Log("FetchAllMedia", "Media is a playlist, we don't need to run a check.")
 		return nil
 	}
 
 	// Check each provider to see if it exists. If not, go query for that media
 	if event.Value.Fields.Apple == nil {
 		// Call Apple Music Query with data
-		log.Println("[FetchAllMedia] Getting media for Apple Music...")
+		logger.Log("FetchAllMedia", "Getting media for Apple Music...")
 	}
 
 	if event.Value.Fields.Spotify == nil {
 		// Call Spotify Query with data
-		log.Println("[FetchAllMedia] Getting media for Spotify...")
+		logger.Log("FetchAllMedia", "Getting media for Spotify...")
 		data, queryErr := querySpotifyMedia(mediaName, mediaCreator, mediaType)
 		if queryErr != nil {
 			logger.LogErr("QuerySpotifyMedia", queryErr, nil)
@@ -62,7 +67,10 @@ func FetchAllMedia(ctx context.Context, event firebase.FirestoreEventSongs) erro
 		var spotifyQueryData spotifyQueryResp
 		json.NewDecoder(*data).Decode(&spotifyQueryData)
 
+		logger.Log("FetchAllMedia", fmt.Sprintf("SpotifyQueryData decoded: %v", spotifyQueryData))
+
 		if len(spotifyQueryData.Tracks.Items) != 0 {
+			logger.Log("FetchAllMedia", fmt.Sprintf("Found tracks: %v", spotifyQueryData.Tracks.Items[0]))
 			// Grab first track item & create song object
 			track := spotifyQueryData.Tracks.Items[0]
 			spotifyData = &map[string]interface{}{
@@ -71,6 +79,7 @@ func FetchAllMedia(ctx context.Context, event firebase.FirestoreEventSongs) erro
 				"url":    track.ExternalURLs.Spotify,
 			}
 		} else if len(spotifyQueryData.Albums.Items) != 0 {
+			logger.Log("FetchAllMedia", fmt.Sprintf("Found albums: %v", spotifyQueryData.Albums.Items[0]))
 			// Grab first track item & create song object
 			album := spotifyQueryData.Albums.Items[0]
 			spotifyData = &map[string]interface{}{
@@ -83,7 +92,7 @@ func FetchAllMedia(ctx context.Context, event firebase.FirestoreEventSongs) erro
 
 	if event.Value.Fields.YouTube == nil {
 		// Call Youtube Query with data
-		log.Println("[FetchAllMedia] Getting media for Youtube Music...")
+		logger.Log("FetchAllMedia", "Getting media for YouTube Music...")
 	}
 
 	// Write data to song document and check if it changed
@@ -107,12 +116,16 @@ func FetchAllMedia(ctx context.Context, event firebase.FirestoreEventSongs) erro
 		return error
 	}
 
+	logger.Log("FetchAllMedia", fmt.Sprintf("Writing blob to Firestore: %v", dataBlob))
+
 	// Write data to db
 	writeDataErr := writeData(dataBlob, docPath[1])
 	if writeDataErr != nil {
 		logger.LogErr("WriteData", writeDataErr, nil)
 		return writeDataErr
 	}
+
+	logger.Log("FetchAllMedia", "Data written successfully.")
 
 	return nil
 }
