+++ cmd/getapplemusicmedia/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.4: Adding more logging
 - v1.0.0-beta.3: Fixing merge
 - v1.0.0-beta.2: Reving Firebase package to the latest
 - v1.0.0-beta.1: Initializing function
