+++ cmd/getapplemusicmedia/helpers.go
@@ -10,6 +10,7 @@ import (
 
 	"cloud.google.com/go/firestore"
 	"github.com/pixelogicdev/gruveebackend/pkg/firebase"
+	"github.com/pixelogicdev/gruveebackend/pkg/sawmill"
 	"github.com/pixelogicdev/gruveebackend/pkg/social"
 )
 
@@ -34,32 +35,48 @@ func initWithEnv() error {
 		return fmt.Errorf("CreateUser [Init Firestore]: %v", err)
 	}
 
-	firestoreClient = client
+	// Initialize Sawmill
+	sawmillLogger, err := sawmill.InitClient(currentProject, os.Getenv("GCLOUD_CONFIG"), os.Getenv("ENVIRONMENT"), "FetchAllMedia")
+	if err != nil {
+		log.Printf("FetchAllMedia [Init Sawmill]: %v", err)
+	}
 
+	firestoreClient = client
+	logger = sawmillLogger
 	return nil
 }
 
 // generateAppleMusicReq created the request object to call Apple Music API
 func generateAppleMusicReq(uri string, method string, devToken string) (*http.Request, error) {
+	logger.Log("GenerateAppleMusicReq", "Starting...")
+
 	// Generate request
 	appleMusicReq, appleMusicReqErr := http.NewRequest(method, uri, nil)
 	if appleMusicReqErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicReqErr)
 	}
 
+	logger.Log("GenerateAppleMusicReq", "Generarted AppleMusicReq")
+
 	// Add headers
 	appleMusicReq.Header.Add("Authorization", "Bearer "+devToken)
+
+	logger.Log("GenerateAppleMusicReq", "Created request successfully.")
 
 	return appleMusicReq, nil
 }
 
 // getAppleMusicTrack will call Apple Music Song API to get the metadata for a song
 func getAppleMusicTrack(trackID string, storefront string, appleDevToken firebase.FirestoreAppleDevJWT) (*firebase.FirestoreMedia, error) {
+	logger.Log("GetAppleMusicTrack", "Starting...")
+
 	// Generate request
 	appleMusicGetTrackReq, appleMusicGetTrackReqErr := generateAppleMusicReq(catalogHostname+"/"+storefront+"/songs/"+trackID, "GET", appleDevToken.Token)
 	if appleMusicGetTrackReqErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicGetTrackReqErr)
 	}
+
+	logger.Log("GetAppleMusicTrack", "Generated Apple Music Request.")
 
 	// Make request
 	getTrackResp, getTrackRespErr := httpClient.Do(appleMusicGetTrackReq)
@@ -67,8 +84,12 @@ func getAppleMusicTrack(trackID string, storefront string, appleDevToken firebas
 		return nil, fmt.Errorf("GetAppleMusicMedia [client.Do]: %v", getTrackRespErr)
 	}
 
+	logger.Log("GetAppleMusicTrack", "Received Track Response.")
+
 	// Check to see if request was valid
 	if getTrackResp.StatusCode != http.StatusOK {
+		logger.Log("GetAppleMusicTrack", "Error received. Decoding body...")
+
 		// Convert Apple Music Error Object
 		var appleMusicErrorObj social.AppleMusicRequestError
 
@@ -80,16 +101,19 @@ func getAppleMusicTrack(trackID string, storefront string, appleDevToken firebas
 		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Track Request]: %v", appleMusicErrorObj)
 	}
 
+	logger.Log("GetAppleMusicTrack", "Successful response.")
+
 	var appleMusicTrackData appleMusicTrackResp
 	respDecodeErr := json.NewDecoder(getTrackResp.Body).Decode(&appleMusicTrackData)
 	if respDecodeErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Response Decoder]: %v", respDecodeErr)
 	}
 
-	log.Println(appleMusicTrackData)
+	logger.Log("GetAppleMusicTrack", "Decoded Apple Musc Track Data.")
 
 	// Check for length to make sure we found a match
 	if len(appleMusicTrackData.Data) == 0 {
+		logger.Log("GetAppleMusicTrack", "No matches found.")
 		return nil, fmt.Errorf("GetAppleMusicMedia [Length Check]: No results in data for id %s", trackID)
 	}
 
@@ -107,24 +131,34 @@ func getAppleMusicTrack(trackID string, storefront string, appleDevToken firebas
 		},
 	}
 
+	logger.Log("GetAppleMusicTrack", "Successfully returning track.")
+
 	return &firestoreMedia, nil
 }
 
 // getAppleMusicPlaylist will call Apple Music Playlist API to get the metadata for a playlist
 func getAppleMusicPlaylist(playlistID string, storefront string, appleDevToken firebase.FirestoreAppleDevJWT) (*firebase.FirestoreMedia, error) {
+	logger.Log("GetAppleMusicPlaylist", "Starting...")
+
 	// Generate request
 	appleMusicGetPlaylistReq, appleMusicGetTrackReqErr := generateAppleMusicReq(catalogHostname+"/"+storefront+"/playlists/"+playlistID, "GET", appleDevToken.Token)
 	if appleMusicGetTrackReqErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicGetTrackReqErr)
 	}
 
+	logger.Log("GetAppleMusicPlaylist", "Generated request.")
+
 	getPlaylistResp, getPlaylistRespErr := httpClient.Do(appleMusicGetPlaylistReq)
 	if getPlaylistRespErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMedia [client.Do]: %v", getPlaylistRespErr)
 	}
 
+	logger.Log("GetAppleMusicPlaylist", "Received playlist response")
+
 	// Check to see if request was valid
 	if getPlaylistResp.StatusCode != http.StatusOK {
+		logger.Log("GetAppleMusicPlaylist", "Error received. Decoding body...")
+
 		// Convert Apple Music Error Object
 		var appleMusicErrorObj social.AppleMusicRequestError
 
@@ -136,13 +170,15 @@ func getAppleMusicPlaylist(playlistID string, storefront string, appleDevToken f
 		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Track Request]: %v", appleMusicErrorObj)
 	}
 
+	logger.Log("GetAppleMusicPlaylist", "Successful response.")
+
 	var appleMusicPlaylistResp appleMusicPlaylistResp
 	respDecodeErr := json.NewDecoder(getPlaylistResp.Body).Decode(&appleMusicPlaylistResp)
 	if respDecodeErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Response Decoder]: %v", respDecodeErr)
 	}
 
-	log.Println(appleMusicPlaylistResp)
+	logger.Log("GetAppleMusicPlaylist", "Successfully decoded response")
 
 	// Check for length to make sure we found a match
 	if len(appleMusicPlaylistResp.Data) == 0 {
@@ -163,28 +199,38 @@ func getAppleMusicPlaylist(playlistID string, storefront string, appleDevToken f
 		},
 	}
 
+	logger.Log("GetAppleMusicPlaylist", "Successfully created playlist data.")
+
 	return &firestoreMedia, nil
 }
 
 // getAppleMusicAlbum will call Apple Music Album API to get the metadata for an album
 func getAppleMusicAlbum(albumID string, storefront string, appleDevToken firebase.FirestoreAppleDevJWT) (*firebase.FirestoreMedia, error) {
+	logger.Log("GetAppleMusicAlbum", "Starting...")
+
 	// Generate request
-	appleMusicGetPlaylistReq, appleMusicGetTrackReqErr := generateAppleMusicReq(catalogHostname+"/"+storefront+"/albums/"+albumID, "GET", appleDevToken.Token)
+	appleMusicGetAlbumReq, appleMusicGetTrackReqErr := generateAppleMusicReq(catalogHostname+"/"+storefront+"/albums/"+albumID, "GET", appleDevToken.Token)
 	if appleMusicGetTrackReqErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicGetTrackReqErr)
 	}
 
-	getPlaylistResp, getPlaylistRespErr := httpClient.Do(appleMusicGetPlaylistReq)
+	logger.Log("GetAppleMusicAlbum", "Generated album request.")
+
+	getAlbumResp, getPlaylistRespErr := httpClient.Do(appleMusicGetAlbumReq)
 	if getPlaylistRespErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMedia [client.Do]: %v", getPlaylistRespErr)
 	}
 
+	logger.Log("GetAppleMusicAlbum", "Received album")
+
 	// Check to see if request was valid
-	if getPlaylistResp.StatusCode != http.StatusOK {
+	if getAlbumResp.StatusCode != http.StatusOK {
+		logger.Log("GetAppleMusicAlbum", "Error received. Decoding body...")
+
 		// Convert Apple Music Error Object
 		var appleMusicErrorObj social.AppleMusicRequestError
 
-		err := json.NewDecoder(getPlaylistResp.Body).Decode(&appleMusicErrorObj)
+		err := json.NewDecoder(getAlbumResp.Body).Decode(&appleMusicErrorObj)
 		if err != nil {
 			return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Request Decoder]: %v", err)
 		}
@@ -193,12 +239,12 @@ func getAppleMusicAlbum(albumID string, storefront string, appleDevToken firebas
 	}
 
 	var appleMusicAlbumResp appleMusicAlbumResp
-	respDecodeErr := json.NewDecoder(getPlaylistResp.Body).Decode(&appleMusicAlbumResp)
+	respDecodeErr := json.NewDecoder(getAlbumResp.Body).Decode(&appleMusicAlbumResp)
 	if respDecodeErr != nil {
 		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Response Decoder]: %v", respDecodeErr)
 	}
 
-	log.Println(appleMusicAlbumResp)
+	logger.Log("GetAppleMusicAlbum", "Decoded response")
 
 	// Check for length to make sure we found a match
 	if len(appleMusicAlbumResp.Data) == 0 {
@@ -218,6 +264,8 @@ func getAppleMusicAlbum(albumID string, storefront string, appleDevToken firebas
 			Images: album.Attributes.Artwork,
 		},
 	}
+
+	logger.Log("GetAppleMusicAlbum", "Successfully created Album response.")
 
 	return &firestoreMedia, nil
 }
