+++ cmd/getspotifymedia/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.6: Adding more logging
 - v1.0.0-beta.5: Fixing merge
 - v1.0.0-beta.4: Reving Firebase package to the latest
 - v1.0.0-beta.3: Moving shared functions to helpers file
