+++ cmd/getspotifymedia/helpers.go
@@ -5,7 +5,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"log"
 	"net/http"
 	"net/url"
 	"os"
@@ -46,11 +45,14 @@ func initWithEnv() error {
 
 // getApiToken checks to see if we have an APIToken for client credential calls
 func getCreds() (*firebase.FirestoreSpotifyAuthToken, error) {
+	logger.Log("GetCreds", "Starting...")
+
 	// Check to see if we have env variables
 	if os.Getenv("SPOTIFY_CLIENTID") == "" || os.Getenv("SPOTIFY_SECRET") == "" {
-		log.Fatalln("GetSpotifyMedia [Check Env Props]: PROPS NOT HERE.")
 		return nil, fmt.Errorf("getSpotifyMedia [Check Env Props]: PROPS NOT HERE")
 	}
+
+	logger.Log("GetCreds", "Found Spotify ClientID in env variables")
 
 	// Check for Auth token in DB
 	authToken, authTokenErr := fetchToken()
@@ -58,17 +60,23 @@ func getCreds() (*firebase.FirestoreSpotifyAuthToken, error) {
 		return nil, authTokenErr
 	}
 
+	logger.Log("GetCreds", "Received token.")
+
 	// Check to see if auth token exists && needs to be refreshed
 	if authToken != nil {
-		log.Println("Checking to see if token needs to be refreshed")
+		logger.Log("GetCreds", "Checking to see if token needs to be refreshed")
 
 		latestCreds, refreshAuthTokenErr := refreshAuthToken(*authToken)
 		if refreshAuthTokenErr != nil {
 			return nil, refreshAuthTokenErr
 		}
 
+		logger.Log("GetCreds", "Auth token refresh complete.")
+
 		return latestCreds, nil
 	}
+
+	logger.Log("GetCreds", "No auth token found.")
 
 	// If we are here, no auth token was found
 	newAuthToken, newAuthTokenErr := generateAuthToken()
@@ -76,23 +84,28 @@ func getCreds() (*firebase.FirestoreSpotifyAuthToken, error) {
 		return nil, newAuthTokenErr
 	}
 
+	logger.Log("GetCreds", "Successfully generated new auth token.")
+
 	// Store new token in DB
 	writeSpotifyAuthErr := writeSpotifyAuthtoken(*newAuthToken)
 	if writeSpotifyAuthErr != nil {
 		return nil, writeSpotifyAuthErr
 	}
 
+	logger.Log("GetCreds", "Successfully stored new auth token in DB.")
+
 	// TheYagich01: "Gejnerated" (08/11/20)
-	log.Println("Generated auth token")
 	return newAuthToken, nil
 }
 
 // fetchToken will grab the Apple Developer Token from DB
 func fetchToken() (*firebase.FirestoreSpotifyAuthToken, error) {
+	logger.Log("FetchToken", "Starting...")
+
 	// Go to Firebase and see if spotifyAuthToken exists
 	snapshot, snapshotErr := firestoreClient.Collection("internal_tokens").Doc("spotifyAuthToken").Get(context.Background())
 	if status.Code(snapshotErr) == codes.NotFound {
-		log.Println("[GetSpotifyMedia] SpotifyAuthToken not found in DB. Need to create.")
+		logger.Log("FetchToken", "SpotifyAuthToken not found in DB. Need to create.")
 		return nil, nil
 	}
 
@@ -100,17 +113,23 @@ func fetchToken() (*firebase.FirestoreSpotifyAuthToken, error) {
 		return nil, fmt.Errorf(snapshotErr.Error())
 	}
 
+	logger.Log("FetchToken", "Successfully got snapshot.")
+
 	var spotifyAuthToken firebase.FirestoreSpotifyAuthToken
 	dataToErr := snapshot.DataTo(&spotifyAuthToken)
 	if dataToErr != nil {
 		return nil, fmt.Errorf(dataToErr.Error())
 	}
 
+	logger.Log("FetchToken", "Successfully decoded auth token.")
+
 	return &spotifyAuthToken, nil
 }
 
 // generateAuthToken will call Spotify creds service to get authToken for making requests
 func generateAuthToken() (*firebase.FirestoreSpotifyAuthToken, error) {
+	logger.Log("GenerateAuthToken", "Starting...")
+
 	// If not there generate new one and store
 	authStr := os.Getenv("SPOTIFY_CLIENTID") + ":" + os.Getenv("SPOTIFY_SECRET")
 
@@ -124,6 +143,8 @@ func generateAuthToken() (*firebase.FirestoreSpotifyAuthToken, error) {
 		return nil, fmt.Errorf(accessTokenReqErr.Error())
 	}
 
+	logger.Log("GenerateAuthToken", "Request generated.")
+
 	accessTokenReq.Header.Add("Content-Type", "application/x-www-form-urlencoded")
 	accessTokenReq.Header.Add("Authorization", "Basic "+base64.StdEncoding.EncodeToString([]byte(authStr)))
 	accessTokenResp, accessTokenRespErr := httpClient.Do(accessTokenReq)
@@ -131,12 +152,16 @@ func generateAuthToken() (*firebase.FirestoreSpotifyAuthToken, error) {
 		return nil, fmt.Errorf(accessTokenRespErr.Error())
 	}
 
+	logger.Log("GenerateAuthToken", "Response received.")
+
 	// Decode the token
 	var spotifyClientCredsAuthResp social.SpotifyClientCredsAuthResp
 	refreshTokenDecodeErr := json.NewDecoder(accessTokenResp.Body).Decode(&spotifyClientCredsAuthResp)
 	if refreshTokenDecodeErr != nil {
 		return nil, fmt.Errorf(refreshTokenDecodeErr.Error())
 	}
+
+	logger.Log("GenerateAuthToken", "Successfully decoded token.")
 
 	// Generate authToken for DB
 	issuedAt := time.Now()
@@ -148,34 +173,45 @@ func generateAuthToken() (*firebase.FirestoreSpotifyAuthToken, error) {
 		Token:     spotifyClientCredsAuthResp.AccessToken,
 	}
 
+	logger.Log("GenerateAuthToken", "Successfully generated token to store in DB.")
+
 	return &authToken, nil
 }
 
 // refreshAuthToken will check to see if Spotify AuthToken needs to be refreshed
 func refreshAuthToken(authToken firebase.FirestoreSpotifyAuthToken) (*firebase.FirestoreSpotifyAuthToken, error) {
+	logger.Log("RefreshAuthToken", "Starting...")
+
 	// Get current time
 	var currentTime = time.Now()
 
-	fmt.Printf("Expires In: %d seconds\n", authToken.ExpiresIn)
-	fmt.Printf("Expires At: %s seconds\n", authToken.ExpiredAt)
+	logger.Log("RefreshAuthToken", fmt.Sprintf("Expires In: %d seconds\n", authToken.ExpiresIn))
+	logger.Log("RefreshAuthToken", fmt.Sprintf("Expires At: %s seconds\n", authToken.ExpiredAt))
 
 	expiredAtTime, expiredAtTimeErr := time.Parse(time.RFC3339, authToken.ExpiredAt)
 	if expiredAtTimeErr != nil {
 		return nil, fmt.Errorf(expiredAtTimeErr.Error())
 	}
 
+	logger.Log("RefreshAuthToken", "Successfully parsed ExpiredAtTime.")
+
 	if currentTime.After(expiredAtTime) {
-		fmt.Println("spotifyAuthToken is expired. Refreshing...")
+		logger.Log("RefreshAuthToken", "SpotifyAuthToken is expired. Refreshing...")
+
 		refreshToken, tokenActionErr := generateAuthToken()
 		if tokenActionErr != nil {
 			return nil, fmt.Errorf(tokenActionErr.Error())
 		}
+
+		logger.Log("RefreshAuthToken", "Successfully generated auth token.")
 
 		// Set new token data in database
 		writeTokenErr := writeSpotifyAuthtoken(*refreshToken)
 		if writeTokenErr != nil {
 			return nil, fmt.Errorf(writeTokenErr.Error())
 		}
+
+		logger.Log("RefreshAuthToken", "Successfully wrote token to Firestore")
 
 		return refreshToken, nil
 	}
@@ -186,10 +222,14 @@ func refreshAuthToken(authToken firebase.FirestoreSpotifyAuthToken) (*firebase.F
 
 // writeSpotifyAuthtoken will take spotifyAuthToken and write it to the internal_tokens collection
 func writeSpotifyAuthtoken(authToken firebase.FirestoreSpotifyAuthToken) error {
+	logger.Log("WriteSpotifyAuthtoken", "Starting...")
+
 	spotifyAuthTokenDoc := firestoreClient.Collection("internal_tokens").Doc("spotifyAuthToken")
 	if spotifyAuthTokenDoc == nil {
 		return fmt.Errorf("spotifyAuthTokenDoc could not be found")
 	}
+
+	logger.Log("WriteSpotifyAuthtoken", "Successfully received SpotifyAuthTokenDoc.")
 
 	// MergeAll doesn't allow custom Go types so we need to create a map
 	authMap := map[string]interface{}{
@@ -204,11 +244,15 @@ func writeSpotifyAuthtoken(authToken firebase.FirestoreSpotifyAuthToken) error {
 		return fmt.Errorf(writeErr.Error())
 	}
 
+	logger.Log("WriteSpotifyAuthtoken", "Successfully set token in Firestore.")
+
 	return nil
 }
 
 // getSpotifyTrack calls Spotify GET track API and converts to Golang Type
 func getSpotifyTrack(trackID string, accessToken string) (*firebase.FirestoreMedia, error) {
+	logger.Log("GetSpotifyTrack", "Starting...")
+
 	// Generate URI
 	spotifyGetURI := spotifyGetTrackURI + "/" + trackID
 
@@ -218,6 +262,8 @@ func getSpotifyTrack(trackID string, accessToken string) (*firebase.FirestoreMed
 		return nil, fmt.Errorf("GetSpotifyMedia [http.NewRequest]: %v", spotifyTrackReqErr)
 	}
 
+	logger.Log("GetSpotifyTrack", "Generated track request.")
+
 	// Add headers and call request
 	spotifyTrackReq.Header.Add("Authorization", "Bearer "+accessToken)
 	getTrackResp, spotifyTrackRespErr := httpClient.Do(spotifyTrackReq)
@@ -225,8 +271,12 @@ func getSpotifyTrack(trackID string, accessToken string) (*firebase.FirestoreMed
 		return nil, fmt.Errorf("GetSpotifyMedia [client.Do]: %v", spotifyTrackRespErr)
 	}
 
+	logger.Log("GetSpotifyTrack", "Received track response.")
+
 	// Check to see if request was valid
 	if getTrackResp.StatusCode != http.StatusOK {
+		logger.Log("GetSpotifyTrack", "Received error. Decoding...")
+
 		// Convert Spotify Error Object
 		var spotifyErrorObj social.SpotifyRequestError
 
@@ -246,6 +296,8 @@ func getSpotifyTrack(trackID string, accessToken string) (*firebase.FirestoreMed
 		return nil, fmt.Errorf("GetSpotifyMedia [Spotify Response Decoder]: %v", respDecodeErr)
 	}
 
+	logger.Log("GetSpotifyTrack", "Successfully decoded response data.")
+
 	// Setup FirestoreMeida object
 	firestoreMedia := firebase.FirestoreMedia{
 		Name:    spotifyTrackData.Name,
@@ -259,11 +311,15 @@ func getSpotifyTrack(trackID string, accessToken string) (*firebase.FirestoreMed
 		},
 	}
 
+	logger.Log("GetSpotifyTrack", "Successfully created firestore media object.")
+
 	return &firestoreMedia, nil
 }
 
 // getSpotifyAlbum calls Spotify API to get playlist metadata
 func getSpotifyPlaylist(playlistID string, accessToken string) (*firebase.FirestoreMedia, error) {
+	logger.Log("GetSpotifyPlaylist", "Starting...")
+
 	// Generate URI
 	spotifyGetURI := spotifyGetPlaylistURI + "/" + playlistID
 
@@ -273,6 +329,8 @@ func getSpotifyPlaylist(playlistID string, accessToken string) (*firebase.Firest
 		return nil, fmt.Errorf("GetSpotifyMedia [http.NewRequest]: %v", spotifyPlaylistReqErr)
 	}
 
+	logger.Log("GetSpotifyPlaylist", "Generated playlist request.")
+
 	// Add headers and call request
 	spotifyPlaylistReq.Header.Add("Authorization", "Bearer "+accessToken)
 	getPlaylistResp, spotifyPlaylistRespErr := httpClient.Do(spotifyPlaylistReq)
@@ -280,8 +338,12 @@ func getSpotifyPlaylist(playlistID string, accessToken string) (*firebase.Firest
 		return nil, fmt.Errorf("GetSpotifyMedia [client.Do]: %v", spotifyPlaylistRespErr)
 	}
 
+	logger.Log("GetSpotifyPlaylist", "Received playlist response.")
+
 	// Check to see if request was valid
 	if getPlaylistResp.StatusCode != http.StatusOK {
+		logger.Log("GetSpotifyPlaylist", "Received error. Decoding...")
+
 		// Convert Spotify Error Object
 		var spotifyErrorObj social.SpotifyRequestError
 
@@ -301,6 +363,8 @@ func getSpotifyPlaylist(playlistID string, accessToken string) (*firebase.Firest
 		return nil, fmt.Errorf("GetSpotifyMedia [Spotify Response Decoder]: %v", respDecodeErr)
 	}
 
+	logger.Log("GetSpotifyPlaylist", "Successfully decoded response.")
+
 	// Setup FirestoreMeida object
 	firestoreMedia := firebase.FirestoreMedia{
 		Name:    playlistData.Name,
@@ -314,11 +378,15 @@ func getSpotifyPlaylist(playlistID string, accessToken string) (*firebase.Firest
 		},
 	}
 
+	logger.Log("GetSpotifyPlaylist", "Successfully created firestore media object.")
+
 	return &firestoreMedia, nil
 }
 
 // getSpotifyAlbum calls Spotify API to get album metadata
 func getSpotifyAlbum(albumID string, accessToken string) (*firebase.FirestoreMedia, error) {
+	logger.Log("GetSpotifyAlbum", "Starting...")
+
 	// Generate URI
 	spotifyGetURI := spotifyGetAlbumURI + "/" + albumID
 
@@ -328,6 +396,8 @@ func getSpotifyAlbum(albumID string, accessToken string) (*firebase.FirestoreMed
 		return nil, fmt.Errorf("GetSpotifyMedia [http.NewRequest]: %v", spotifyAlbumReqErr)
 	}
 
+	logger.Log("GetSpotifyAlbum", "Generated album request.")
+
 	// Add headers and call request
 	spotifyAlbumReq.Header.Add("Authorization", "Bearer "+accessToken)
 	spotifyAlbumResp, spotifyAlbumRespErr := httpClient.Do(spotifyAlbumReq)
@@ -335,8 +405,12 @@ func getSpotifyAlbum(albumID string, accessToken string) (*firebase.FirestoreMed
 		return nil, fmt.Errorf("GetSpotifyMedia [client.Do]: %v", spotifyAlbumRespErr)
 	}
 
+	logger.Log("GetSpotifyAlbum", "Received album response.")
+
 	// Check to see if request was valid
 	if spotifyAlbumResp.StatusCode != http.StatusOK {
+		logger.Log("GetSpotifyAlbum", "Received error. Decoding...")
+
 		// Convert Spotify Error Object
 		var spotifyErrorObj social.SpotifyRequestError
 
@@ -356,6 +430,8 @@ func getSpotifyAlbum(albumID string, accessToken string) (*firebase.FirestoreMed
 		return nil, fmt.Errorf("GetSpotifyMedia [Spotify Response Decoder]: %v", respDecodeErr)
 	}
 
+	logger.Log("GetSpotifyAlbum", "Successfully decoded album data.")
+
 	// Setup FirestoreMeida object
 	firestoreMedia := firebase.FirestoreMedia{
 		Name:    albumData.Name,
@@ -369,16 +445,22 @@ func getSpotifyAlbum(albumID string, accessToken string) (*firebase.FirestoreMed
 		},
 	}
 
+	logger.Log("GetSpotifyAlbum", "Successfully created firestore media object.")
+
 	return &firestoreMedia, nil
 }
 
 // generateArtistsString takes in a list of artists and returns a comma separated string
 func generateArtistsString(artists []spotifyArtist) string {
+	logger.Log("GenerateArtistsString", "Starting...")
+
 	var creators = []string{}
 
 	for _, artist := range artists {
 		creators = append(creators, artist.Name)
 	}
+
+	logger.Log("GenerateArtistsString", "Successfully appended artists.")
 
 	return strings.Join(creators, ", ")
 }
