+++ cmd/tokengen/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.6: Adding more logging & adding helpers file
 - v1.0.0-beta.5: Fixing merge
 - v1.0.0-beta.4: Reving firebase and social package to latest
 - v1.0.0-beta.3: Changing enviornment variable check location
