+++ cmd/updatealgolia/.version
@@ -1,3 +1,4 @@
+- v1.0.0-beta.11: Adding more logging & adding types and helpers files
 - v1.0.0-beta.10: Fixing merge
 - v1.0.0-beta.9: Reving Firebase version
 - v1.0.0-beta.8: Adding DisplayName to payload
