+++ pkg/sawmill/.version
@@ -1,1 +1,2 @@
+- v1.0.0-beta.2: Updated Log function param order
 - v1.0.0-beta.1: Created Sawmill
